moved helper structs to types.rs
