+++ src/send/mod.rs
@@ -1,11 +1,16 @@
-use std::sync::{Arc, Mutex};
+pub mod v2;
+
+use std::io::Cursor;
+use std::sync::Arc;
 
 pub use payjoin::send::RequestBuilder as PdkRequestBuilder;
 
 use crate::error::PayjoinError;
 use crate::transaction::PartiallySignedTransaction;
-use crate::uri::{Uri, Url};
-use crate::FeeRate;
+use crate::uri::Uri;
+
+use crate::send::v2::ContextV2;
+use crate::types::{Amount, FeeRate, Request};
 
 ///Builder for sender-side payjoin parameters
 ///
@@ -18,11 +23,7 @@ impl From<PdkRequestBuilder<'static>> for RequestBuilder {
     }
 }
 
-// impl <'a>From<PdkRequestBuilder<'a>> for RequestBuilder {
-//     fn from(value: PdkRequestBuilder<'a>) -> Self {
-//             Self(value)
-//     }
-// }
+
 
 impl RequestBuilder {
     /// Prepare an HTTP request and request context to process the response
@@ -64,6 +65,46 @@ impl RequestBuilder {
             Err(e) => Err(e.into()),
         }
     }
+    /// Offer the receiver contribution to pay for his input.
+    ///
+    /// These parameters will allow the receiver to take `max_fee_contribution` from given change
+    /// output to pay for additional inputs. The recommended fee is `size_of_one_input * fee_rate`.
+    ///
+    /// `change_index` specifies which output can be used to pay fee. If `None` is provided, then
+    /// the output is auto-detected unless the supplied transaction has more than two outputs.
+    ///
+    /// `clamp_fee_contribution` decreases fee contribution instead of erroring.
+    ///
+    /// If this option is true and a transaction with change amount lower than fee
+    /// contribution is provided then instead of returning error the fee contribution will
+    /// be just lowered in the request to match the change amount.
+    pub fn build_with_additional_fee(
+        &self,
+        max_fee_contribution: Arc<Amount>,
+        change_index: Option<u8>,
+        min_fee_rate: Arc<FeeRate>,
+        clamp_fee_contribution: bool,
+    ) -> Result<Arc<RequestContext>, PayjoinError> {
+        match self.0.clone().build_with_additional_fee(
+            (*max_fee_contribution).clone().into(),
+            change_index.map(|x| x as usize),
+            (*min_fee_rate).into(),
+            clamp_fee_contribution,
+        ) {
+            Ok(e) => Ok(Arc::new(e.into())),
+            Err(e) => Err(e.into()),
+        }
+    }
+    /// Perform Payjoin without incentivizing the payee to cooperate.
+    ///
+    /// While it's generally better to offer some contribution some users may wish not to.
+    /// This function disables contribution.
+    pub fn build_non_incentivizing(&self) -> Result<Arc<RequestContext>, PayjoinError> {
+        match self.0.clone().build_non_incentivizing() {
+            Ok(e) => Ok(Arc::new(e.into())),
+            Err(e) => Err(e.into()),
+        }
+    }
 }
 pub struct RequestContext(payjoin::send::RequestContext);
 
@@ -73,64 +114,59 @@ impl From<payjoin::send::RequestContext> for RequestContext {
     }
 }
 
+#[derive(Clone)]
+pub struct RequestContextV1 {
+    pub request: Request,
+    pub context_v1: Arc<ContextV1>,
+}
+
+#[derive(Clone)]
+pub struct RequestContextV2 {
+    pub request: Request,
+    pub context_v2: Arc<ContextV2>,
+}
+
+impl RequestContext {
+    /// Extract serialized V1 Request and Context from a Payjoin Proposal
+    pub fn extract_v1(&self) -> Result<RequestContextV1, PayjoinError> {
+        match self.0.clone().extract_v1() {
+            Ok(e) => Ok(RequestContextV1 { request: e.0.into(), context_v1: Arc::new(e.1.into()) }),
+            Err(e) => Err(e.into()),
+        }
+    }
+    ///Extract serialized Request and Context from a Payjoin Proposal.
+    /// In order to support polling, this may need to be called many times to be encrypted with new unique nonces to make independent OHTTP requests.
+    /// The ohttp_proxy merely passes the encrypted payload to the ohttp gateway of the receiver
+    pub fn extract_v2(&self,  ohttp_proxy_url:String ) -> Result<RequestContextV2, PayjoinError> {
+        match self.0.clone().extract_v2(ohttp_proxy_url.as_str()) {
+            Ok(e) => Ok(RequestContextV2 { request: e.0.into(), context_v2: Arc::new(e.1.into()) }),
+            Err(e) => Err(e.into()),
+        }
+    }
+
+}
 ///Data required for validation of response.
-
-///This type is used to process the response. It is returned from PjUriExt::create_pj_request() method and you only need to call .process_response() on it to continue BIP78 flow.
-
-pub struct Context(Mutex<Option<payjoin::send::ContextV1>>);
-
-impl Context {
-    fn _get_context(&self) -> Option<payjoin::send::ContextV1> {
-        let mut data_guard = self.0.lock().unwrap();
-        Option::take(&mut *data_guard)
-    }
-}
-
-impl From<payjoin::send::ContextV1> for Context {
+/// This type is used to process the response. Get it from RequestBuilder's build methods. Then you only need to call .process_response() on it to continue BIP78 flow.
+pub struct ContextV1(payjoin::send::ContextV1);
+impl From<payjoin::send::ContextV1> for ContextV1 {
     fn from(value: payjoin::send::ContextV1) -> Self {
-        Self(Mutex::new(Some(value)))
+        Self(value)
     }
 }
 
-///Represents data that needs to be transmitted to the receiver.
-
-///You need to send this request over HTTP(S) to the receiver.
-#[derive(Clone, Debug)]
-pub struct Request {
-    ///URL to send the request to.
-    ///
-    ///This is full URL with scheme etc - you can pass it right to reqwest or a similar library.
-    pub url: Arc<Url>,
-    ///Bytes to be sent to the receiver.
-    ///
-    ///This is properly encoded PSBT, already in base64. You only need to make sure Content-Type is text/plain and Content-Length is body.len() (most libraries do the latter automatically).
-    pub body: Vec<u8>,
-}
-
-impl From<payjoin::receive::v2::Request> for Request {
-    fn from(value: payjoin::receive::v2::Request) -> Self {
-        Self { url: Arc::new(value.url.into()), body: value.body }
-    }
-}
-impl From<payjoin::send::Request> for Request {
-    fn from(value: payjoin::send::Request) -> Self {
-        Self { url: Arc::new(value.url.into()), body: value.body }
+impl ContextV1 {
+    ///Decodes and validates the response.
+    /// Call this method with response from receiver to continue BIP78 flow. If the response is valid you will get appropriate PSBT that you should sign and broadcast.
+    pub fn process_response(
+        &self,
+        response: Vec<u8>,
+    ) -> Result<Arc<PartiallySignedTransaction>, PayjoinError> {
+        let mut decoder = Cursor::new(response);
+        match self.0.clone().process_response(&mut decoder) {
+            Ok(e) => Ok(Arc::new(e.into())),
+            Err(e) => Err(e.into()),
+        }
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use crate::PartiallySignedTransaction;
 
-    #[test]
-    fn official_vectors() {
-        let original = "cHNidP8BAHECAAAAAVuDh6O7xLpvJm70AWI6N25VtXzMiknZxAwcPtGoB/VHAAAAAAD+////AuYPECQBAAAAFgAUHAMjFjcTerY7Cmi4se8VqWIW5HgA4fUFAAAAABYAFL6Az084ngrVLQfpl3hccYjeF+EQAAAAAAABAIQCAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/////wNSAQH/////AgDyBSoBAAAAFgAUQ0p5pXSyKswNZuFoJjltIQhFnpkAAAAAAAAAACZqJKohqe3i9hw/cdHe/T+pmd+jaVN1XGkGiXmZYrSL69g2l06M+QAAAAABAR8A8gUqAQAAABYAFENKeaV0sirMDWbhaCY5bSEIRZ6ZAQhrAkcwRAIgFe1S3DHoDIPHogsTU/9UD7IqPbNXDYfyU2JZT9HKD7oCIAxU7sZzUcoGsmM3lDetos/3N5fM5oynmzuvsrFiILN5ASECfoMstJPrqnyhems+r158wTniKIBaPkkCinDC4VdvmsYAIgIDXVq5OYL7D4Ur28OTJ77j0lZrSPzO5XGmkL/KIF7wKmgQgTP32gAAAIABAACAAQAAgAAA";
-
-        let original_psbt = PartiallySignedTransaction::from_string(original.to_string()).unwrap();
-        eprintln!("original: {:#?}", original_psbt);
-
-        let pj_uri_string = "BITCOIN:BCRT1Q7WXQ0R2JHJKX8HQS3SHLKFAEAEF38C38SYKGZY?amount=1&pj=https://example.comOriginal".to_string();
-        let _uri = crate::Uri::new(pj_uri_string).unwrap();
-        eprintln!("address: {:#?}", _uri.address().as_string());
-    }
-}
